pyproject.toml
[tool.poetry]
name = "hw-2-13"
version = "0.1.0"
description = ""
authors = ["Vitalii Bielimov <vbit2000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
setuptools = "^69.2.0"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"
asyncpg = "^0.29.0"
libgravatar = "^1.0.4"
pydantic = {extras = ["email"], version = "^2.6.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
